Merger SertifikatkjedevalidatorTester for for produksjon og tester
- Drar også ut lesing av sertifikater til SertifikatUtility
