+++ Difi.Felles.Utility.Tester/Utilities/SertifikatUtility.cs
@@ -1,6 +1,5 @@
 ﻿using System.Security.Cryptography.X509Certificates;
 using ApiClientShared;
-using ApiClientShared.Enums;
 
 namespace Difi.Felles.Utility.Tester.Utilities
 {
@@ -11,6 +10,11 @@ namespace Difi.Felles.Utility.Tester.Utilities
         public static X509Certificate2 GetProduksjonsMottakerSertifikatOppslagstjenesten()
         {
             return new X509Certificate2(ResourceUtility.ReadAllBytes(true, "Prod", "produksjonsmottakersertifikatFraOppslagstjenesten.pem"));
+        }
+
+        public static X509Certificate2 GetFunksjoneltTestmiljøMottakerSertifikatOppslagstjenesten()
+        {
+            return new X509Certificate2(ResourceUtility.ReadAllBytes(true, "Test", "testmottakersertifikatFraOppslagstjenesten.pem"));
         }
 
         public static X509Certificate2 NotActivatedTestCertificate()
@@ -26,6 +30,11 @@ namespace Difi.Felles.Utility.Tester.Utilities
         public static X509Certificate2 TestIntegrasjonssertifikat()
         {
             return GetPostenCertificate();
+        }
+
+        public static X509Certificate2 GetEnhetstesterSelvsignertSertifikat()
+        {
+            return new X509Certificate2(ResourceUtility.ReadAllBytes(true, "Enhetstester", "difi-enhetstester.cer"));
         }
 
         public static X509Certificate2 GetPostenCertificate()
