+++ Difi.Felles.Utility/Sertifikatvalidator.cs
